+++ lib/fluent_ui.dart
@@ -56,6 +56,7 @@ export 'src/controls/navigation/tab_view.dart';
 export 'src/controls/surfaces/calendar/calendar_view.dart';
 export 'src/controls/surfaces/bottom_sheet.dart';
 export 'src/controls/surfaces/dialog.dart';
+export 'src/controls/surfaces/expander.dart';
 export 'src/controls/surfaces/flyout/flyout.dart';
 export 'src/controls/surfaces/info_bar.dart';
 export 'src/controls/surfaces/list_tile.dart';
@@ -65,6 +66,7 @@ export 'src/controls/surfaces/tooltip.dart';
 
 export 'src/controls/utils/divider.dart';
 export 'src/controls/utils/hover_button.dart';
+export 'src/controls/utils/info_badge.dart';
 export 'src/controls/utils/scrollbar.dart';
 
 export 'src/controls/form/auto_suggest_box.dart';