+++ lib/src/controls/navigation/navigation_view/pane.dart
@@ -54,285 +54,6 @@ enum PaneDisplayMode {
   auto,
 }
 
-class NavigationPaneItem with Diagnosticable {
-  final GlobalKey itemKey = GlobalKey();
-
-  NavigationPaneItem();
-}
-
-/// The item used by [NavigationView] to display the tiles.
-///
-/// On [PaneDisplayMode.compact], only [icon] is displayed, and [title] is
-/// used as a tooltip. On the other display modes, [icon] and [title] are
-/// displayed in a [Row].
-///
-/// This is the only [NavigationPaneItem] that is affected by [NavigationIndicator]s
-///
-/// See also:
-///
-///   * [PaneItemSeparator], used to group navigation items
-///   * [PaneItemHeader], used to label groups of items.
-class PaneItem extends NavigationPaneItem {
-  /// Creates a pane item.
-  PaneItem({
-    required this.icon,
-    this.title,
-    this.focusNode,
-    this.autofocus = false,
-  });
-
-  /// The title used by this item. If the display mode is top
-  /// or compact, this is shown as a tooltip. If it's open, this
-  /// is shown by the side of the [icon].
-  ///
-  /// The text style is fetched from the closest [NavigationPaneThemeData]
-  ///
-  /// If this is a [Text], its [Text.data] is used to display the tooltip. The
-  /// tooltip is only displayed only on compact mode and when the item is not
-  /// disabled.
-  /// It is also used by [Semantics] to allow screen readers to
-  /// read the screen.
-  ///
-  /// Usually a [Text] widget.
-  final Widget? title;
-
-  /// The icon used by this item.
-  ///
-  /// Usually an [Icon] widget
-  final Widget icon;
-
-  /// {@macro flutter.widgets.Focus.focusNode}
-  final FocusNode? focusNode;
-
-  /// {@macro flutter.widgets.Focus.autofocus}
-  final bool autofocus;
-
-  /// Used to construct the pane items all around [NavigationView]. You can
-  /// customize how the pane items should look like by overriding this method
-  Widget build(
-    BuildContext context,
-    bool selected,
-    VoidCallback? onPressed, {
-    PaneDisplayMode? displayMode,
-    bool showTextOnTop = true,
-    bool? autofocus,
-  }) {
-    final PaneDisplayMode mode = displayMode ??
-        _NavigationBody.maybeOf(context)?.displayMode ??
-        PaneDisplayMode.minimal;
-    assert(displayMode != PaneDisplayMode.auto);
-    final bool isTop = mode == PaneDisplayMode.top;
-    final bool isCompact = mode == PaneDisplayMode.compact;
-    final bool isOpen =
-        [PaneDisplayMode.open, PaneDisplayMode.minimal].contains(mode);
-    final NavigationPaneThemeData theme = NavigationPaneTheme.of(context);
-
-    final String titleText =
-        title != null && title is Text ? (title! as Text).data ?? '' : '';
-
-    return Container(
-      key: itemKey,
-      height: !isTop ? 36.0 : null,
-      width: isCompact ? _kCompactNavigationPanelWidth : null,
-      margin: const EdgeInsets.only(right: 6.0, left: 6.0, bottom: 4.0),
-      alignment: Alignment.center,
-      child: HoverButton(
-        autofocus: autofocus ?? this.autofocus,
-        focusNode: focusNode,
-        onPressed: onPressed,
-        cursor: theme.cursor,
-        builder: (context, states) {
-          final textStyle = selected
-              ? theme.selectedTextStyle?.resolve(states)
-              : theme.unselectedTextStyle?.resolve(states);
-          final textResult = titleText.isNotEmpty
-              ? Padding(
-                  padding: theme.labelPadding ?? EdgeInsets.zero,
-                  child: Text(titleText, style: textStyle),
-                )
-              : const SizedBox.shrink();
-          Widget child = Flex(
-            direction: isTop ? Axis.vertical : Axis.horizontal,
-            textDirection: isTop ? ui.TextDirection.ltr : ui.TextDirection.rtl,
-            mainAxisAlignment: isTop || !isOpen
-                ? MainAxisAlignment.center
-                : MainAxisAlignment.end,
-            children: [
-              if (isOpen) Expanded(child: textResult),
-              () {
-                final icon = Padding(
-                  padding: theme.iconPadding ?? EdgeInsets.zero,
-                  child: IconTheme.merge(
-                    data: IconThemeData(
-                      color: (selected
-                              ? theme.selectedIconColor?.resolve(states)
-                              : theme.unselectedIconColor?.resolve(states)) ??
-                          textStyle?.color,
-                      size: 16.0,
-                    ),
-                    child: Center(child: this.icon),
-                  ),
-                );
-                if (isOpen) return icon;
-                return icon;
-              }(),
-            ],
-          );
-          if (isTop && showTextOnTop) {
-            child = Row(mainAxisSize: MainAxisSize.min, children: [
-              child,
-              textResult,
-            ]);
-          }
-          child = AnimatedContainer(
-            duration: theme.animationDuration ?? Duration.zero,
-            curve: theme.animationCurve ?? standartCurve,
-            decoration: BoxDecoration(
-              color: () {
-                final ButtonState<Color?> tileColor = theme.tileColor ??
-                    ButtonState.resolveWith((states) {
-                      if (isTop) return Colors.transparent;
-                      return ButtonThemeData.uncheckedInputColor(
-                        FluentTheme.of(context),
-                        states,
-                      );
-                    });
-                final newStates = states.toSet()..remove(ButtonStates.disabled);
-                return tileColor.resolve(
-                  selected ? {ButtonStates.hovering} : newStates,
-                );
-              }(),
-              borderRadius: BorderRadius.circular(4.0),
-            ),
-            child: child,
-          );
-          child = Semantics(
-            label: title == null ? null : titleText,
-            selected: selected,
-            child: FocusBorder(
-              child: child,
-              focused: states.isFocused,
-              renderOutside: false,
-            ),
-          );
-          if (((isTop && !showTextOnTop) || isCompact) &&
-              titleText.isNotEmpty &&
-              !states.isDisabled) {
-            return Tooltip(
-              message: titleText,
-              style: TooltipThemeData(
-                textStyle: title is Text ? (title as Text).style : null,
-              ),
-              child: child,
-            );
-          }
-          return child;
-        },
-      ),
-    );
-  }
-}
-
-/// Separators for grouping navigation items. Set the color property to
-/// [Colors.transparent] to render the separator as space. Uses a [Divider]
-/// under the hood, consequently uses the closest [DividerThemeData].
-///
-/// See also:
-///   * [PaneItem], the item used by [NavigationView] to render tiles
-///   * [PaneItemHeader], used to label groups of items.
-class PaneItemSeparator extends NavigationPaneItem {
-  /// Creates an item separator.
-  PaneItemSeparator({this.color, this.thickness});
-
-  /// The color used by the [Divider].
-  final Color? color;
-
-  /// The separator thickness. Defaults to 1.0
-  final double? thickness;
-
-  Widget build(BuildContext context, Axis direction) {
-    return Divider(
-      key: itemKey,
-      direction: direction,
-      style: DividerThemeData(
-        thickness: thickness,
-        decoration: color != null ? BoxDecoration(color: color) : null,
-        verticalMargin: const EdgeInsets.symmetric(
-          horizontal: 8.0,
-          vertical: 10.0,
-        ),
-        horizontalMargin: const EdgeInsets.symmetric(
-          horizontal: 8.0,
-          vertical: 10.0,
-        ),
-      ),
-    );
-  }
-}
-
-/// Headers for labeling groups of items. This is not displayed if the display
-/// mode is [PaneDisplayMode.compact]
-///
-/// See also:
-///   * [PaneItem], the item used by [NavigationView] to render tiles
-///   * [PaneItemSeparator], used to group navigation items
-class PaneItemHeader extends NavigationPaneItem {
-  /// Creates a pane header.
-  PaneItemHeader({required this.header});
-
-  /// The header. The default style is [NavigationPaneThemeData.itemHeaderTextStyle],
-  /// but can be overriten by [Text.style].
-  ///
-  /// Usually a [Text] widget.
-  final Widget header;
-
-  Widget build(BuildContext context) {
-    assert(debugCheckHasFluentTheme(context));
-    final theme = NavigationPaneTheme.of(context);
-    return Padding(
-      key: itemKey,
-      padding: theme.iconPadding ?? EdgeInsets.zero,
-      child: DefaultTextStyle(
-        style: theme.itemHeaderTextStyle ?? const TextStyle(),
-        child: header,
-        softWrap: false,
-        maxLines: 1,
-        overflow: TextOverflow.fade,
-        textAlign: TextAlign.left,
-      ),
-    );
-  }
-}
-
-extension _ItemsExtension on List<NavigationPaneItem> {
-  /// Get the all the item offets in this list
-  List<Offset> getPaneItemsOffsets(GlobalKey<State<StatefulWidget>> paneKey) {
-    return map((e) {
-      // Gets the item global position
-      final itemContext = e.itemKey.currentContext;
-      if (itemContext == null) return Offset.zero;
-      final box = itemContext.findRenderObject()! as RenderBox;
-      final globalPosition = box.localToGlobal(Offset.zero);
-      // And then convert it to the local position
-      final paneContext = paneKey.currentContext;
-      if (paneContext == null) return Offset.zero;
-      final paneBox = paneKey.currentContext!.findRenderObject() as RenderBox;
-      final position = paneBox.globalToLocal(globalPosition);
-      return position;
-    }).toList();
-  }
-
-  /// Get all the item sizes in this list
-  List<Size> getPaneItemsSizes() {
-    return map((e) {
-      final context = e.itemKey.currentContext;
-      if (context == null) return Size.zero;
-      final box = context.findRenderObject()! as RenderBox;
-      return box.size;
-    }).toList();
-  }
-}
-
 /// The pane used by [NavigationView].
 ///
 /// The [NavigationView] doesn't perform any navigation tasks automatically.
@@ -350,12 +71,14 @@ class NavigationPane with Diagnosticable {
     this.key,
     this.selected,
     this.onChanged,
+    this.size,
     this.header,
     this.items = const [],
     this.footerItems = const [],
     this.autoSuggestBox,
     this.autoSuggestBoxReplacement,
     this.displayMode = PaneDisplayMode.auto,
+    this.customPane,
     this.menuButton,
     this.scrollController,
     this.indicatorBuilder = defaultNavigationIndicator,
@@ -369,9 +92,14 @@ class NavigationPane with Diagnosticable {
   /// [PaneDisplayMode.auto] is used by default.
   final PaneDisplayMode displayMode;
 
+  final NavigationPaneWidget? customPane;
+
   /// The menu button used by this pane. If null and [onDisplayModeRequested]
   /// is null
   final Widget? menuButton;
+
+  /// The size of the pane in its various mode.
+  final NavigationPaneSize? size;
 
   /// The header of the pane.
   ///
@@ -441,23 +169,22 @@ class NavigationPane with Diagnosticable {
 
   static Widget defaultNavigationIndicator({
     required BuildContext context,
-    int? index,
-    required List<Offset> Function() offsets,
-    required List<Size> Function() sizes,
-    required Axis axis,
+    required NavigationPane pane,
+    Axis? axis,
     required Widget child,
   }) {
-    if (index == null) return child;
+    if (pane.selected == null) return child;
     assert(debugCheckHasFluentTheme(context));
     final theme = NavigationPaneTheme.of(context);
+    axis ??= Axis.horizontal;
 
     final left = theme.iconPadding?.left ?? theme.labelPadding?.left ?? 0;
     final right = theme.labelPadding?.right ?? theme.iconPadding?.right ?? 0;
 
     return StickyNavigationIndicator(
-      index: index,
-      offsets: offsets,
-      sizes: sizes,
+      index: pane.selected!,
+      offsets: () => pane.effectiveItems.getPaneItemsOffsets(pane.paneKey),
+      sizes: pane.effectiveItems.getPaneItemsSizes,
       child: child,
       color: theme.highlightColor,
       curve: theme.animationCurve ?? Curves.linear,
@@ -487,7 +214,8 @@ class NavigationPane with Diagnosticable {
   }
 
   List<NavigationPaneItem> get effectiveItems {
-    return (allItems..removeWhere((i) => i is! PaneItem));
+    return (allItems
+      ..removeWhere((i) => i is! PaneItem || i is PaneItemAction));
   }
 
   /// Check if the provided [item] is selected on not.
@@ -509,7 +237,7 @@ class NavigationPane with Diagnosticable {
   }) {
     if (pane.menuButton != null) return pane.menuButton!;
     return Container(
-      width: _kCompactNavigationPanelWidth,
+      width: pane.size?.compactWidth ?? _kCompactNavigationPanelWidth,
       margin: padding,
       child: PaneItem(
         title: itemTitle,
@@ -524,6 +252,101 @@ class NavigationPane with Diagnosticable {
   }
 }
 
+/// Configure the size of the pane in its various mode.
+///
+/// ```dart
+/// NavigationView(
+///   pane: NavigationPane(
+///     size: NavigationPaneSize(
+///       openWidth: MediaQuery.of(context).size.width / 5,
+///       openMinWidth: 250,
+///       openMaxWidth: 320,
+///     ),
+///   ),
+/// )
+/// ```
+///
+/// See also:
+///
+///  * [NavigationPane], which this configures the size of
+///  * [NavigationView], used to display [NavigationPane]s
+class NavigationPaneSize {
+  /// The height of the pane when he is in top mode.
+  ///
+  /// If the value is null, [kOneLineTileHeight] is used.
+  final double? topHeight;
+
+  /// The width of the pane when he is in compact mode.
+  ///
+  /// If the value is null, [_kCompactNavigationPanelWidth] is used.
+  final double? compactWidth;
+
+  /// The width of the pane when he is open.
+  ///
+  /// If the value is null, [_kOpenNavigationPanelWidth] is used.
+  /// The width can be based on MediaQuery and used
+  /// with [minWidth] and [maxWidth].
+  final double? openWidth;
+
+  /// The minimum width of the pane when he is open.
+  ///
+  /// If width is smaller than minWidth, minWidth is used as width.
+  /// minWidth must be smaller or equal to maxWidth.
+  final double? openMinWidth;
+
+  /// The maximum width of the pane when he is open.
+  ///
+  /// If width is greater than maxWidth, maxWidth is used as width.
+  /// maxWidth must be greater or equal than minWidth.
+  final double? openMaxWidth;
+
+  const NavigationPaneSize({
+    this.topHeight,
+    this.compactWidth,
+    this.openWidth,
+    this.openMinWidth,
+    this.openMaxWidth,
+  }) : assert(
+          openMinWidth == null ||
+              openMaxWidth == null ||
+              openMinWidth <= openMaxWidth,
+          'openMinWidth should be greater than openMaxWidth',
+        );
+}
+
+class NavigationPaneWidgetData {
+  const NavigationPaneWidgetData({
+    required this.content,
+    required this.appBar,
+    required this.scrollController,
+    required this.paneKey,
+    required this.listKey,
+    required this.pane,
+  });
+
+  final Widget content;
+  final Widget appBar;
+  final ScrollController scrollController;
+  final Key? paneKey;
+  final GlobalKey? listKey;
+  final NavigationPane pane;
+}
+
+/// Base class for creating custom navigation panes.
+///
+/// ```dart
+/// class CustomNavigationPane extends NavigationPaneWidget {
+///   CustomNavigationPane();
+///
+///   @override
+///   Widget build(BuildContext context, NavigationPaneWidgetData data) {
+///   }
+/// }
+/// ```
+abstract class NavigationPaneWidget {
+  Widget build(BuildContext context, NavigationPaneWidgetData data);
+}
+
 /// Creates a top navigation pane.
 ///
 /// ![Top Pane Anatomy](https://docs.microsoft.com/en-us/windows/uwp/design/controls-and-patterns/images/navview-pane-anatomy-horizontal.png)
@@ -531,11 +354,13 @@ class _TopNavigationPane extends StatelessWidget {
   _TopNavigationPane({
     required this.pane,
     this.listKey,
+    this.scrollbarKey,
     this.appBar,
   }) : super(key: pane.key);
 
   final NavigationPane pane;
   final GlobalKey? listKey;
+  final GlobalKey? scrollbarKey;
   final NavigationAppBar? appBar;
 
   Widget _buildItem(BuildContext context, NavigationPaneItem item) {
@@ -564,12 +389,10 @@ class _TopNavigationPane extends StatelessWidget {
   Widget build(BuildContext context) {
     assert(debugCheckHasFluentTheme(context));
     Widget topBar = SizedBox(
-      height: kOneLineTileHeight,
+      height: pane.size?.topHeight ?? kOneLineTileHeight,
       child: pane.indicatorBuilder(
         context: context,
-        index: pane.selected,
-        offsets: () => pane.effectiveItems.getPaneItemsOffsets(pane.paneKey),
-        sizes: pane.effectiveItems.getPaneItemsSizes,
+        pane: pane,
         axis: Axis.vertical,
         child: Row(key: pane.paneKey, children: [
           Expanded(
@@ -584,6 +407,8 @@ class _TopNavigationPane extends StatelessWidget {
                 ),
               Expanded(
                 child: Scrollbar(
+                  key: scrollbarKey,
+                  controller: pane.scrollController,
                   isAlwaysShown: false,
                   // A single child scroll view is used instead of a ListView
                   // because the Row implies the cross axis alignment to center,
@@ -626,12 +451,14 @@ class _CompactNavigationPane extends StatelessWidget {
     required this.pane,
     this.paneKey,
     this.listKey,
+    this.scrollbarKey,
     this.onToggle,
   }) : super(key: pane.key);
 
   final NavigationPane pane;
   final Key? paneKey;
   final GlobalKey? listKey;
+  final GlobalKey? scrollbarKey;
   final VoidCallback? onToggle;
 
   Widget _buildItem(BuildContext context, NavigationPaneItem item) {
@@ -668,13 +495,10 @@ class _CompactNavigationPane extends StatelessWidget {
       key: paneKey,
       duration: theme.animationDuration ?? Duration.zero,
       curve: theme.animationCurve ?? Curves.linear,
-      width: _kCompactNavigationPanelWidth,
+      width: pane.size?.compactWidth ?? _kCompactNavigationPanelWidth,
       child: pane.indicatorBuilder(
         context: context,
-        index: pane.selected,
-        offsets: () => pane.effectiveItems.getPaneItemsOffsets(pane.paneKey),
-        sizes: pane.effectiveItems.getPaneItemsSizes,
-        axis: Axis.horizontal,
+        pane: pane,
         child: Align(
           key: pane.paneKey,
           alignment: Alignment.topCenter,
@@ -711,6 +535,8 @@ class _CompactNavigationPane extends StatelessWidget {
               ),
             Expanded(
               child: Scrollbar(
+                key: scrollbarKey,
+                controller: pane.scrollController,
                 isAlwaysShown: false,
                 child: ListView(key: listKey, primary: true, children: [
                   ...pane.items.map((item) {
@@ -734,6 +560,7 @@ class _OpenNavigationPane extends StatelessWidget {
     required this.pane,
     this.paneKey,
     this.listKey,
+    this.scrollbarKey,
     this.onToggle,
     this.onItemSelected,
   }) : super(key: pane.key);
@@ -741,6 +568,7 @@ class _OpenNavigationPane extends StatelessWidget {
   final NavigationPane pane;
   final Key? paneKey;
   final GlobalKey? listKey;
+  final GlobalKey? scrollbarKey;
   final VoidCallback? onToggle;
   final VoidCallback? onItemSelected;
 
@@ -792,17 +620,24 @@ class _OpenNavigationPane extends StatelessWidget {
       }
       return const SizedBox.shrink();
     }();
+    var paneWidth = pane.size?.openWidth ?? _kOpenNavigationPanelWidth;
+    if (pane.size?.openMaxWidth != null &&
+        paneWidth > pane.size!.openMaxWidth!) {
+      paneWidth = pane.size!.openMaxWidth!;
+    }
+    if (pane.size?.openMinWidth != null &&
+        paneWidth < pane.size!.openMinWidth!) {
+      paneWidth = pane.size!.openMinWidth!;
+    }
+
     return AnimatedContainer(
       key: paneKey,
       duration: theme.animationDuration ?? Duration.zero,
       curve: theme.animationCurve ?? Curves.linear,
-      width: _kOpenNavigationPanelWidth,
+      width: paneWidth,
       child: pane.indicatorBuilder(
         context: context,
-        index: pane.selected,
-        offsets: () => pane.effectiveItems.getPaneItemsOffsets(pane.paneKey),
-        sizes: pane.effectiveItems.getPaneItemsSizes,
-        axis: Axis.horizontal,
+        pane: pane,
         child: Column(key: pane.paneKey, children: [
           Container(
             margin: pane.autoSuggestBox != null ? EdgeInsets.zero : topPadding,
@@ -833,6 +668,8 @@ class _OpenNavigationPane extends StatelessWidget {
             ),
           Expanded(
             child: Scrollbar(
+              key: scrollbarKey,
+              controller: pane.scrollController,
               isAlwaysShown: false,
               child: ListView(key: listKey, primary: true, children: [
                 ...pane.items.map((item) {